requirements.txt
pyaudio
openai
pvporcupine
boto3
pygame
wave
uuid
picovoice==2.2.1
python-dotenv